Added persistence storage for chart
