+++ package.json
@@ -249,6 +249,7 @@
     "electron-debug": "3.1.0",
     "electron-localshortcut": "^3.2.1",
     "electron-log": "4.2.4",
+    "electron-store": "^8.0.0",
     "electron-updater": "4.3.9",
     "electronbar": "^2.0.0",
     "history": "5.0.0",
