+++ src/package.json
@@ -1,7 +1,7 @@
 {
   "name": "binocular",
   "productName": "Binocular",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "An opinionated Binance client for crypto trading. GPL-3.0-or-later",
   "main": "./main.prod.js",
   "author": {
